Refactor (#60)
* Change website models.py

Updating models to represented changes after mentor talk. I will
include a client registration page and thus the models needs to
totally change. #51

* Modify website models and add allauth

Changin website models and adding allauth library to handle
the creation of new users for client. This will allo users
to make and manage their own bookings. #51 #52

* Change website admin.py to reflect model changes

Update admin.py to reflect the changes in the model. #56

* Update views.py to reflect new model

Updating views.py to reflect the changes in the new model and
structure of the program. #54

* Update forms.py to reflect new model

Updating forms.py to reflect the new model. #57

* Refactor models.py to reflect new changes

Updating models.py to reflect the new approach for reservations.#51

* Modify model to reflect new site workflow

Removing the client model and use the built in User class
to identify the user logged. #51

* Change URL to implement new views

Adding new views and updating the naming convention in URLS #54

* Add new view to display client reservation

Adding new view to display to clients their upcoming reservation
this will allow them to change that reservation or book a new one.#54

* Updating forms.py to capture the logged user

Changing forms.py to reflect new changes in model and workflow
the form will capture the logged user to associate the reservation
when the forms is submitted. #57

* Add new template for client reservations

Adding new template to display upcoming reservations to client
and allow them to view current and future reservatiosn and
modified them. #54

* Add validation to avoid double booking

Adding validation to prevent double reservations to happen. The
validation check user logged, time and date. #55

* Update requirements.txt

* Update URLs to reflect new functionalties

Updating URL to reflect the new functionalties in the app like
double reservation validation. #58

* Fix delete_reservation to fit the new schema

Fixing URL for delete_reservation and adjust the view to reflect
new URL. #54

* Update naming in search_reservation

Update naming convention in search_reservation to reflect new
naming convention. #54

* Style allauth template to match site

Override the default template to bootstrapping them to match the
site style and made them responsive. #53

* Include allauth links to base.html navbar

Adding allauth links to navbar to expand functionality. #40

* Change phone number time input to tel

Changing phone number field type to Tel to help with input validation
The validation will display now when a phone number is not correct.
If phone number is not in Irish national formar will ask for a new
number or a international number. #57

* Create function to validate duplicate reservations

Move validation to a separate function to be used in make reservations
and update reservations. #59

* Add further style to tempplates

Adding further style and adjusted template variables for
client_reservations, confirmed, update_reservation templates.
Tracked in #54

* Add migrations files to git